Fixes Collision Bug
This bug would cause some collisions to be missed, resulting in the player passing through obstacles; the fix was to adjust the Rigidbody of the Player and the Obstacle Prefab to be "continuous" rather than "discrete". Also adjusted the Fixed Timestamp in the Project Settings > Time to 0.01 rather than 0.02 to update collisions more frequently.